requirements.txt
aiogram==3.13.1
flask==3.0.3
requests==2.32.3
aiofiles==23.2.1
aiohttp==3.10.5
python-dotenv==1.0.1
